+++ Sources/PorscheConnect/CommonResponseObjects/GenericValue.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public struct GenericValue: Codable {
+  public let value: Double
+  public let unit: String
+  public let unitTranslationKey: String
+  public let unitTranslationKeyV2: String
+}
