+++ Tests/PorscheConnectTests/Formatters/GenericValueFormatterTests.swift
@@ -1,0 +1,68 @@
+import XCTest
+
+@testable import PorscheConnect
+
+final class GenericValueFormatterTests: XCTestCase {
+
+  var formatter: GenericValueFormatter!
+  override func setUp() {
+    formatter = GenericValueFormatter()
+    formatter.timeZone = .init(abbreviation: "GMT")!
+    formatter.defaultDate = DateComponents(year: 2023, month: 1, day: 24, hour: 0, minute: 0, second: 0).date
+  }
+
+  override func tearDown() {
+    formatter = nil
+  }
+
+  // MARK: - Defaults
+
+  func testDefaultLocaleIsCurrent() {
+    XCTAssertEqual(formatter.locale, .current)
+  }
+
+  // MARK: - Specific locales
+
+  private let typicalBatteryPercentage = GenericValue(
+    value: 12,
+    unit: "PERCENT",
+    unitTranslationKey: "GRAY_SLICE_UNIT_PERCENT",
+    unitTranslationKeyV2: "TC.UNIT.PERCENT"
+  )
+  private let typicalDaysRemaining = GenericValue(
+    value: -710,
+    unit: "DAYS",
+    unitTranslationKey: "GRAY_SLICE_UNIT_DAY",
+    unitTranslationKeyV2: "TC.UNIT.DAYS"
+  )
+
+  func testAmericanEnglishFormatting() {
+    formatter.locale = Locale(identifier: "en_US")
+    XCTAssertEqual(formatter.string(from: typicalBatteryPercentage, scalar: 0.01), "12%")
+    XCTAssertEqual(formatter.string(from: typicalDaysRemaining, scalar: -1), "Jan 4, 2025")
+  }
+
+  func testCanadianEnglishFormatting() {
+    formatter.locale = Locale(identifier: "en_CA")
+    XCTAssertEqual(formatter.string(from: typicalBatteryPercentage, scalar: 0.01), "12%")
+    XCTAssertEqual(formatter.string(from: typicalDaysRemaining, scalar: -1), "Jan 4, 2025")
+  }
+
+  func testUnitedKingdomEnglishFormatting() {
+    formatter.locale = Locale(identifier: "en_GB")
+    XCTAssertEqual(formatter.string(from: typicalBatteryPercentage, scalar: 0.01), "12%")
+    XCTAssertEqual(formatter.string(from: typicalDaysRemaining, scalar: -1), "4 Jan 2025")
+  }
+
+  func testChineseFormatting() {
+    formatter.locale = Locale(identifier: "zh_CN")
+    XCTAssertEqual(formatter.string(from: typicalBatteryPercentage, scalar: 0.01), "12%")
+    XCTAssertEqual(formatter.string(from: typicalDaysRemaining, scalar: -1), "2025年1月4日")
+  }
+
+  func testGermanFormatting() {
+    formatter.locale = Locale(identifier: "de_DE")
+    XCTAssertEqual(formatter.string(from: typicalBatteryPercentage, scalar: 0.01), "12 %")
+    XCTAssertEqual(formatter.string(from: typicalDaysRemaining, scalar: -1), "04.01.2025")
+  }
+}
